+++ R/InteractivePlots.R
@@ -172,7 +172,7 @@ InteractivePlots <- R6::R6Class(
                 try({
                   shiny::tagList(
                     shiny::h3(name),
-                    TreatmentPatterns::createSankeyDiagram2(
+                    TreatmentPatterns::createSankeyDiagram(
                       treatmentPathways = data$treatmentPathways %>%
                         dplyr::filter(.data$db == name)
                     )